+++ src/main/java/com/example/solidconnection/auth/client/AppleOAuthClient.java
@@ -2,7 +2,7 @@ package com.example.solidconnection.auth.client;
 
 import com.example.solidconnection.auth.dto.oauth.AppleTokenDto;
 import com.example.solidconnection.auth.dto.oauth.AppleUserInfoDto;
-import com.example.solidconnection.config.client.AppleOAuthClientProperties;
+import com.example.solidconnection.auth.client.config.AppleOAuthClientProperties;
 import com.example.solidconnection.common.exception.CustomException;
 import io.jsonwebtoken.Jwts;
 import lombok.RequiredArgsConstructor;
