+++ src/main/java/com/example/solidconnection/security/provider/SiteUserAuthenticationProvider.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.security.provider;
 
-import com.example.solidconnection.config.security.JwtProperties;
+import com.example.solidconnection.security.config.JwtProperties;
 import com.example.solidconnection.security.userdetails.SiteUserDetails;
 import com.example.solidconnection.security.userdetails.SiteUserDetailsService;
 import com.example.solidconnection.security.authentication.JwtAuthentication;
