+++ src/test/java/com/example/solidconnection/security/provider/SiteUserAuthenticationProviderTest.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.security.provider;
 
-import com.example.solidconnection.config.security.JwtProperties;
+import com.example.solidconnection.security.config.JwtProperties;
 import com.example.solidconnection.common.exception.CustomException;
 import com.example.solidconnection.security.authentication.SiteUserAuthentication;
 import com.example.solidconnection.security.userdetails.SiteUserDetails;
